pyproject.toml
[project]
name = "skill-bridge-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi==0.115.0",
    "httpx==0.27.2",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.9.2",
    "pytest>=9.0.3",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "sqlalchemy==2.0.35",
    "uvicorn==0.30.6",
]
